+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ## Unreleased
 
 - 💥 BREAKING: Move `node-sass` to `peerDependencies`
+- 🐛 BUGFIX: Do not require `node-sass` if `@example scss` is not used
 
 
 ## 2.1.0: 2018-04-02